requirements.txt
streamlit
opencv-python
numpy
torch
torchvision
facenet-pytorch
scikit-learn
pandas
pillow
pyttsx3
streamlit-webrtc
av
pytz
gTTS
fastapi
uvicorn
python-multipart
SQLAlchemy
pydantic
passlib
python-dotenv
face-recognition

